refactor: LadderSymbol을 view로 패키지 이동
