+++ src/main/java/view/LadderSymbol.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import java.util.Arrays;
