+++ src/test/java/domain/LadderSymbolTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import view.LadderSymbol;
+
 class LadderSymbolTest {
 
     @ParameterizedTest
